TIP-1073: move ValueFactory
